python/requirements.txt
playwright==1.57.0
python-dotenv==1.0.1
